requirements.txt
fastapi>=0.115
uvicorn>=0.29
pronouncing>=0.2.0
pyttsx3>=2.90
python-dotenv>=1.0
pydub>=0.25.1
torchaudio==2.2.0
torchaudio>=2.2.0
transformers>=4.37.4
gtts>=2.3.2
SQLAlchemy>=2.0.32
psycopg2-binary>=2.9.7
alembic>=1.11.1
pydantic>=2.8.0